Fixed Python scripts execution
